fix: cantidad_ventas Int -> Long

+++ app/src/main/java/org/luisito/gestor360/data/repository/InventarioRepository.kt
@@ -73,7 +73,7 @@ class InventarioRepository(private val context: Context = AppContextHolder.conte
         val totales = TotalesVentas(
             efectivo = ventasHoy.sumOf { it.efectivo },
             transferencia = ventasHoy.sumOf { it.transferencia },
-            cantidad_ventas = ventasHoy.size
+            cantidad_ventas = ventasHoy.size.toLong()
         )
         
         // Productos modificados hoy
